requirements.txt
playwright==1.58.0
requests==2.33.0
requests-oauthlib==2.0.0
PyYAML==6.0.2
google-auth==2.41.1
python-dateutil==2.9.0.post0
